model-server/requirements.txt
flask==3.0.2
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.6.1
xgboost==2.1.4
catboost==1.2.7
h5py==3.12.1

flask-cors==3.0.10
tensorflow==2.17.1
pillow==10.2.0
numpy>=1.26.0
python-dotenv==0.19.0

librosa
opencv-python
matplotlib
keras
jupyter
notebook
tqdm
openai
resampy
